% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_fric_ashape.R
\name{fa_fric_ashape}
\alias{fa_fric_ashape}
\title{Compute Functional Richness (FRic) with
\ifelse{html}{\out{&alpha;}}{\eqn{$\alpha$}{alpha}}-shape}
\usage{
fa_fric_ashape(traits, sp_com, stand = FALSE, avalue = "auto")
}
\arguments{
\item{traits}{The matrix dataset for which you want to compute the index}

\item{sp_com}{Site-species matrix with sites as rows and species as columns
if not provided, the function considers all species with equal
abundance in a single site. This can be either a \code{matrix},
a \code{data.frame}, or a \code{\link[Matrix:Matrix]{Matrix::Matrix()}} object.}

\item{stand}{a boolean indicating whether to standardize FRic values over
the observed maximum over all species (default: \code{FALSE}).
This scales FRic between 0 and 1.
\strong{NB}: The maximum FRic values only considers species that are
present in \strong{both} site-species and trait matrices.
If you want to consider species that are absent
in the site-species matrix, add corresponding columns of 0s.}

\item{avalue}{The value of the
\ifelse{html}{\out{&alpha;}}{\eqn{$\alpha$}{alpha}} parameter to compute the
\ifelse{html}{\out{&alpha;}}{\eqn{$\alpha$}{alpha}}-shape.
Set to \code{"auto"} (the default) to get the
\ifelse{html}{\out{&alpha;*}}{\eqn{$\alpha^*$}{alpha*}} value as defined in
Gruson (2020).}
}
\value{
a data.frame with two columns:
\itemize{
\item \code{site} the names of the sites as the row names of the input \code{sp_com},
\item \code{FRic} the values of functional richness at each site.
}
}
\description{
Functional Richness is computed as the volume of the
\ifelse{html}{\out{&alpha;}}{\eqn{$\alpha$}{alpha}}-shape from all
included traits.
}
\section{Parallelization}{


The computation of this function can be parallelized thanks to
\code{\link[future:plan]{future::plan()}}. To get more information on how to parallelize your
computation please refer to the parallelization vignette with:
\code{vignette("parallel", package = "fundiversity")}

}

\examples{
data(traits_birds, package = "fundiversity")
fa_fric_ashape(traits_birds[,-1])
}
\references{
Gruson H. 2020. Estimation of colour volumes as concave hypervolumes using
\ifelse{html}{\out{&alpha;}}{\eqn{$\alpha$}{alpha}}-shapes. Methods in
Ecology and Evolution, early view \doi{10.1111/2041-210X.13398}
}
